requirements.txt
streamlit>=1.20
pandas>=1.3
plotly>=5.0
python-docx>=0.8.11
pymupdf>=1.22.0
scikit-learn>=1.0.0
openai>=0.27.0